Consolidate Being-H0.5 policy modules
